+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # CHANGELOG
 
+## 1.3.0
+
+- Removed deprecated wmic.exe usage in windows and use powershell instead
+
 ## 1.2.0
 
-- [#24] Improve performance 
-- [#27] Make tests deterministic 
+- [#24] Improve performance
+- [#27] Make tests deterministic
 - [#29] Improve CI configurations
 
 ## 1.1.1